+++ src/App.css
@@ -7,6 +7,7 @@
 body {
   background-color: #303f47;
   color: #96FF7D;
+  padding-bottom: 2em;
 }
